IT WORKS!
This deployment works in a basic way, there is some weirdness with the kafka deployment and also topics being created in Kafdrop. For some reason, it cannot respond to topics created in Kafdrop (maybe?), not really sure. A kafka thing I think. But the parallel processing is working, although the partition assignment in kafka is odd.